Rename apply_constants.rs
